split up localize
